Удалил лишний модуль js, оформил вывод массива в виде ф-ии и подключил
