+++ js/mock-data.js
@@ -1,3 +1,5 @@
+import {getRandomIntInclusive, getRandom} from './util.js';
+
 // Число сгенерированных объектов:
 const COUNT_FOTO_CARDS = 25;
 
@@ -19,4 +21,36 @@ const AUTHORS = [
   'Тапок Убийца'
 ];
 
-export {COUNT_FOTO_CARDS, MESAGGES, AUTHORS};
+// Ф-я создания массива-объекта коментариев:
+function getComments () {
+  const comment = [];
+  for(let j = 1; j < getRandomIntInclusive(1, 6); j++) {
+    comment.push({
+      id: j,
+      avatar: `img/avatar-${ j }.svg`,
+      message: getRandom(MESAGGES),
+      name: getRandom(AUTHORS),
+    });
+  }
+  return comment;
+}
+
+// Ф-я создания объекта фотокарточки
+function createCard(i) {
+  return {
+    id: i,
+    url: `photos/${  i  }.jpg`,
+    description: 'Какое то фото..',
+    likes: getRandomIntInclusive(15, 200),
+    comments: getComments(),
+  };
+}
+// Создания массива из 25 объектов фотокарточки:
+const mocks = function getMockPhotos(){
+  const photoElements = [];
+  for(let i = 1; i <= COUNT_FOTO_CARDS; i++){
+    photoElements.push(createCard(i));
+  }
+  return photoElements;
+};
+export {mocks};
